- move `MaterialTheme` to `-m2` module (update default API which now requires colors provided)
