+++ compose-desktop/src/main/kotlin/main.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.application
-import com.mikepenz.markdown.compose.Markdown
+import com.mikepenz.markdown.m2.Markdown
 
 fun main() = application {
     Window(onCloseRequest = ::exitApplication, title = "Markdown Sample") {
