+++ multiplatform-markdown-renderer-m2/.gitignore
@@ -1,0 +1,1 @@
+/build
