+++ multiplatform-markdown-renderer-m2/build.gradle.kts
@@ -1,6 +1,5 @@
 import com.vanniktech.maven.publish.SonatypeHost
 import org.jetbrains.kotlin.gradle.dsl.KotlinVersion
-import org.jetbrains.kotlin.gradle.dsl.KotlinVersion.KOTLIN_1_8
 
 plugins {
     kotlin("multiplatform")
@@ -11,7 +10,7 @@ plugins {
 }
 
 android {
-    namespace = "com.mikepenz.markdown"
+    namespace = "com.mikepenz.markdown.m2"
     compileSdk = Versions.androidCompileSdk
 
     defaultConfig {
@@ -35,21 +34,6 @@ android {
 
     tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
         kotlinOptions.jvmTarget = "11"
-
-        kotlinOptions {
-            if (project.findProperty("composeCompilerReports") == "true") {
-                freeCompilerArgs += listOf(
-                    "-P",
-                    "plugin:androidx.compose.compiler.plugins.kotlin:reportsDestination=${project.buildDir.absolutePath}/compose_compiler"
-                )
-            }
-            if (project.findProperty("composeCompilerMetrics") == "true") {
-                freeCompilerArgs += listOf(
-                    "-P",
-                    "plugin:androidx.compose.compiler.plugins.kotlin:metricsDestination=${project.buildDir.absolutePath}/compose_compiler"
-                )
-            }
-        }
     }
 }
 
@@ -72,12 +56,6 @@ kotlin {
         compilations {
             all {
                 kotlinOptions.jvmTarget = "11"
-            }
-        }
-
-        testRuns["test"].executionTask.configure {
-            useJUnit {
-                excludeCategories("org.intellij.markdown.ParserPerformanceTest")
             }
         }
     }
@@ -133,14 +111,14 @@ kotlin {
 }
 
 dependencies {
+    commonMainApi(project(":multiplatform-markdown-renderer"))
+
     commonMainApi(Deps.Markdown.core)
 
     commonMainCompileOnly(compose.runtime)
-    commonMainCompileOnly(compose.ui)
-    commonMainCompileOnly(compose.foundation)
+    // commonMainCompileOnly(compose.ui)
+    // commonMainCompileOnly(compose.foundation)
     commonMainCompileOnly(compose.material)
-
-    "androidMainImplementation"(Deps.Compose.coilCompose)
 }
 
 tasks.dokkaHtml.configure {
