+++ multiplatform-markdown-renderer-m2/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Multiplatform Markdown Renderer
+POM_DESCRIPTION=Kotlin Multiplatform Markdown Renderer. (Android, Desktop, ...) powered by Compose Multiplatform
+POM_ARTIFACT_ID=multiplatform-markdown-renderer-m2
