+++ multiplatform-markdown-renderer-m2/src/commonMain/kotlin/com/mikepenz/markdown/m2/Markdown.kt
@@ -1,0 +1,31 @@
+package com.mikepenz.markdown.m2
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.mikepenz.markdown.compose.components.MarkdownComponents
+import com.mikepenz.markdown.compose.components.markdownComponents
+import com.mikepenz.markdown.model.*
+import org.intellij.markdown.flavours.MarkdownFlavourDescriptor
+import org.intellij.markdown.flavours.gfm.GFMFlavourDescriptor
+
+@Composable
+fun Markdown(
+    content: String,
+    colors: MarkdownColors = markdownColor(),
+    typography: MarkdownTypography = markdownTypography(),
+    padding: MarkdownPadding = markdownPadding(),
+    modifier: Modifier = Modifier.fillMaxSize(),
+    flavour: MarkdownFlavourDescriptor = GFMFlavourDescriptor(),
+    imageTransformer: ImageTransformer = ImageTransformerImpl(),
+    components: MarkdownComponents = markdownComponents(),
+) = com.mikepenz.markdown.compose.Markdown(
+    content,
+    colors,
+    typography,
+    padding,
+    modifier,
+    flavour,
+    imageTransformer,
+    components,
+)
