+++ multiplatform-markdown-renderer/src/commonMain/kotlin/com/mikepenz/markdown/model/MarkdownColors.kt
@@ -1,7 +1,5 @@
 package com.mikepenz.markdown.model
 
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
 import androidx.compose.ui.graphics.Color
@@ -28,7 +26,7 @@ interface MarkdownColors {
 }
 
 @Immutable
-private class DefaultMarkdownColors(
+class DefaultMarkdownColors(
     override val text: Color,
     override val codeText: Color,
     override val linkText: Color,
@@ -36,20 +34,3 @@ private class DefaultMarkdownColors(
     override val inlineCodeBackground: Color,
     override val dividerColor: Color,
 ) : MarkdownColors
-
-@Composable
-fun markdownColor(
-    text: Color = MaterialTheme.colors.onBackground,
-    codeText: Color = MaterialTheme.colors.onBackground,
-    linkText: Color = text,
-    codeBackground: Color = MaterialTheme.colors.onBackground.copy(alpha = 0.1f),
-    inlineCodeBackground: Color = codeBackground,
-    dividerColor: Color = MaterialTheme.colors.onSurface.copy(alpha = 0.12f),
-): MarkdownColors = DefaultMarkdownColors(
-    text = text,
-    codeText = codeText,
-    linkText = linkText,
-    codeBackground = codeBackground,
-    inlineCodeBackground = inlineCodeBackground,
-    dividerColor = dividerColor,
-)
