+++ settings.gradle.kts
@@ -9,5 +9,7 @@ pluginManagement {
 rootProject.name = "multiplatform-markdown-renderer-root"
 
 include(":multiplatform-markdown-renderer")
+include(":multiplatform-markdown-renderer-m2")
+
 include(":app")
 include(":compose-desktop")
